Prevent redundant hit of treasuremap.php
If the user is not on any treasure map, the quest
QuestRelicHunter.default_map_id will be null. Hitting the treasure map
endpoint with null as the map_id returns an error in the
messageData.popup response.

It's not shown to the user but it's still
best to not do extraneous network requests to HG servers.